refactor: timeVoteDto를 record로 변경
