+++ backend/src/main/java/middle_point_search/backend/domains/timeVoteRoom/dto/response/FindTimeVoteRoomResponse.java
@@ -1,0 +1,13 @@
+package middle_point_search.backend.domains.timeVoteRoom.dto.response;
+
+import java.time.LocalDate;
+import java.util.List;
+
+public record FindTimeVoteRoomResponse(
+	Boolean existence,
+	List<LocalDate> dates
+) {
+	public static FindTimeVoteRoomResponse from(Boolean existence, List<LocalDate> dates) {
+		return new FindTimeVoteRoomResponse(existence, dates);
+	}
+}
